requirements.txt
absl-py==0.7.1
addict==2.2.1
altgraph==0.16.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
astunparse==1.6.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
boto==2.49.0
boto3==1.14.12
botocore==1.17.12
Brlapi==0.6.6
cachetools==4.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro==1.0.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
easydict==1.9
efficientnet==1.1.0
entrypoints==0.3
evdev==0.7.0
feedparser==5.2.1
future==0.17.1
gast==0.3.3
gensim==3.8.3
gevent==1.4.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.21.1
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
imageio==2.5.0
imutils==0.5.2
iotop==0.6
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightgbm==2.3.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mmcv-full==1.1.1
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.4
nibabel==2.5.0
notebook==5.7.8
numpy==1.16.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
opt-einsum==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pefile==2019.4.18
pexpect==4.2.1
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==12.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
Pygments==2.4.2
pygobject==3.26.1
PyInstaller==3.5
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyreadline==2.1
pyRFC3339==1.0
pyrsistent==0.15.2
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.0
python-debian==0.1.32
python-slugify==4.0.0
pytz==2019.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rsa==4.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.12.0
sklearn==0.0
smart-open==2.0.0
soupsieve==1.9.4
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-io==0.15.0
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
text-unidecode==1.3
tifffile==2020.6.3
torch==1.6.0
torchvision==0.7.0
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.2
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xdiagnose==3.8.8
xgboost==0.90
xkit==0.0.0
yapf==0.30.0
zope.interface==4.3.2
